Extract inline-examples.js - Keep it clean!
