+++ src/stories/index.js
@@ -1,1 +1,2 @@
 import './Calendar-stories';
+import './inline-examples';
